Extract _os.py from _legacy.py (#1416)
