+++ modules/helpers/__init__.py
@@ -41,4 +41,5 @@ from ._file_utils import *  # noqa: F403
 from ._settings import *  # noqa: F403
 from ._git import *  # noqa: F403
 from ._forms import *  # noqa: F403
+from ._os import *  # noqa: F403
 from ._paths import *  # noqa: F403
